Use bigdecimal instead of double for price
[#168074457]

+++ src/main/kotlin/com/boclips/terry/application/orders/converters/metadataConverters/PriceFieldConverter.kt
@@ -2,6 +2,7 @@ package com.boclips.terry.application.orders.converters.metadataConverters
 
 import com.boclips.terry.domain.model.Price
 import java.lang.NumberFormatException
+import java.math.BigDecimal
 import java.util.Currency
 
 object PriceFieldConverter {
@@ -23,11 +24,11 @@ object PriceFieldConverter {
                 .takeIf { it }
                 ?.let { unparsedPrice.replace(delimiterToCurrencyPair.key, "") }
                 ?.let { it.toDoubleOrNull() }
-                ?.let { return Price.WithCurrency(it, currency = delimiterToCurrencyPair.value) }
+                ?.let { return Price.WithCurrency(BigDecimal.valueOf(it), currency = delimiterToCurrencyPair.value) }
         }
 
 
-        return unparsedPrice.toDoubleOrNull()?.let { Price.WithoutCurrency(it) }
+        return unparsedPrice.toDoubleOrNull()?.let { Price.WithoutCurrency(BigDecimal.valueOf(it)) }
             ?: Price.InvalidPrice
     }
 }
