+++ src/main/kotlin/com/boclips/terry/domain/model/Price.kt
@@ -1,9 +1,10 @@
 package com.boclips.terry.domain.model
 
+import java.math.BigDecimal
 import java.util.Currency
 
-sealed class Price(val value: Double) {
-    class WithCurrency(value: Double, val currency: Currency) : Price(value)
-    class WithoutCurrency(value: Double) : Price(value)
-    object InvalidPrice : Price(-1.0)
+sealed class Price(val value: BigDecimal) {
+    class WithCurrency(value: BigDecimal, val currency: Currency) : Price(value)
+    class WithoutCurrency(value: BigDecimal) : Price(value)
+    object InvalidPrice : Price(BigDecimal.valueOf(-1.0))
 }
